requirements.txt
watchdog
pyyaml
docopt
wakeonlan
pythonping
plexapi
requests
